fixed #1
