+++ bin/sol-profiler.js
@@ -3,7 +3,7 @@
 const parser= require("solparse"),
      clc    = require("cli-color"),
      table  = require('table'),
-     utils  = require('./../utils');
+     {data, file, profile}  = require('./../utils');
 
 let config  = {
     border: {
@@ -42,28 +42,24 @@ async function generateReport(path){
 
         // Adding filename and solidity version
         let version;
-
-        let pragma = await utils.getPragma(path);
-        let code = await utils.processFile(path);
+        let pragma = await file.getPragma(path);
+        let code = await file.process(path);
         let source = parser.parse(pragma + '\n\n' + code);
         /* jshint ignore:end */
         if(source.body[0].type == 'PragmaStatement')
             version = source.body[0].start_version.version;
-        
         let fileArray = path.split('/');
-        let file = fileArray[fileArray.length -1];
-        let contractName = file.substr(0, file.length - 4);
-        tableRows.push(['',clc.greenBright("File: " + file + 
-                " , Solidity Pragma: " + version), '','','','']);
+        let fileName = fileArray[fileArray.length -1];
+        let contractName = fileName.substr(0, fileName.length - 4);
+        tableRows.push(['',clc.greenBright("File: " + fileName + " , Solidity Pragma: " + version), '','','','']);
 
         // Adding header row 
         tableRows.push([clc.whiteBright.bold('Contract/Library'), clc.whiteBright.bold('Function/Constructor'), clc.whiteBright.bold('Visibility'), clc.whiteBright.bold('View/Pure'), clc.whiteBright.bold('Returns'), clc.whiteBright.bold('Modifiers')]);
-
         source.body.forEach(function(contract) {
             if(contract.type == 'ContractStatement' || contract.type == 'LibraryStatement') {
                 contract.body.forEach(function(part) {
                 if(part.type == 'ConstructorDeclaration' || (part.type == 'FunctionDeclaration' && part.is_abstract == false)) {
-                    let {contractName, functionName, visibility, viewOrPure, returns, modifiers} = parsePartData(contract, part);
+                    let {contractName, functionName, visibility, viewOrPure, returns, modifiers} = data.parseData(contract, part);
                     tableRows.push([contractName, functionName, visibility, viewOrPure, returns, modifiers]);
                 }
             });
@@ -71,94 +67,16 @@ async function generateReport(path){
         });
         /* jshint ignore:start */
         var tableData = table.table(tableRows, config); /* jshint ignore:end */
-        let fileData = tableData.replace(
-            /[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g, ''); // clearing color formatting
-        require("fs").writeFileSync(contractName + "_Profile.txt", fileData);
-        console.log(tableData); 
+
+        // Store profile in profiles folder
+        profile.store(tableData, contractName);
+        
+        //Render profile on console
+        console.log(tableData);
     }catch(error){
         console.log(clc.red(error.message));
     };
 
-}
-
-function parsePartData(contract, part) {
-    let contractName = clc.cyanBright(contract.name);
-    if(contract.type == 'LibraryStatement')
-        contractName = contractName + clc.white(' -lib');
-    
-    let funcName = null;
-    if(part.type == 'ConstructorDeclaration')
-        funcName = 'constructor';
-        else if(part.type == 'FunctionDeclaration'){
-                funcName = part.name || '';
-        }
-
-    let params = [];
-    if(part.params) {
-        part.params.forEach(function(param) {
-            if(param.storage_location)
-                params.push(param.literal.literal + ' ' + clc.cyan(param.storage_location));
-            else
-                params.push(param.literal.literal);
-        });
-    funcName += '(' + params.join(',') + ')';
-    }
-    else {
-        //Check fallback
-        if(!funcName && !part.name && !part.params && !part.returnParams)
-            funcName = '()' + clc.white(' -fallback');
-        else
-            funcName += '()';
-    }
-
-    // Default is public
-    let visibility = clc.magentaBright("public");
-    let  viewOrPure = '';
-    let  returns    = [];
-    let  custom     = [];
-
-    if(part.modifiers) {
-        part.modifiers.forEach(function(mod) {
-            switch(mod.name) {
-                case "public":
-                    break;
-                case "private":
-                    visibility = clc.redBright("private");
-                    break;
-                case "internal":
-                    visibility = clc.red("internal");
-                    break;
-                case "external":
-                    visibility = clc.magenta("external");
-                    break;
-                case "view":
-                    viewOrPure = clc.yellow("view");
-                    break;
-                case "pure":
-                    viewOrPure = clc.yellowBright("pure");
-                    break;
-                default:
-                    custom.push(mod.name);
-            }
-        });
-    }
-    if(part.returnParams) {
-        part.returnParams.params.forEach(function(param) {
-            if(param.storage_location)
-                returns.push(param.literal.literal + ' ' + clc.cyan(param.storage_location));
-            else
-                returns.push(param.literal.literal);
-        });
-    }
-
-    return {
-        contractName:   contractName,
-        functionName:   clc.blueBright(funcName),
-        visibility  :   visibility,
-        viewOrPure  :   viewOrPure,
-        returns     :   clc.white(returns),
-        modifiers   :   clc.white(custom)
-    };
 }
 
 module.exports = generateReport(filePath);
